pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gobblet-rl"
version = "1.1.0"
authors = [
  { name="Elliot Tower", email="elliot@elliottower.com" },
]
description = "Multi-Agent Reinforcement Learning Environment for the board game Gobblet using PettingZoo."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies=[
    "gym==0.26.1",
    "gymnasium==0.27.1",
    "numpy==1.23.5",
    "pettingzo==1.22.3",
    "setuptools>=65.6.3",
    "tianshou==0.4.11",
    "torch==1.12.1",
    "pytest==7.1.2",
    "pygame==2.1.2",
    "ray==2.2.0"

]

[project.urls]
"Homepage" = "https://github.com/elliottower/gobblet-rl/"
"Bug Tracker" = "https://github.com/elliottower/gobblet/issues"